fibonacci rabbit

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,18 +7,18 @@
 
 int main(void)
 {
-	float f1, f2, fn;
+	long double f1, f2, fn;
 	int c;
 
 	f1 = 1;
 	f2 = 2;
 
-	printf("%.0f, %.0f", f1, f2);
+	printf("%.0Lf, %.0Lf", f1, f2);
 
 	for (c = 3; c <= 98; c++)
 	{
 		fn = (f1 + f2);
-		printf(", %.0f", fn);
+		printf(", %.0Lf", fn);
 
 		f1 = (f2);
 		f2 = (fn);
